Extract logic for preview breakpoints to a separate hook.
